requirements.txt
pandas
streamlit
fasttext
fasttext-langdetect
jupyter
notebook
ipykernel
ipywidgets
tqdm
watchdog
streamlit-option-menu
stqdm
openpyxl
fastapi
pandas 
uvicorn 
xlsxwriter 
opencc 
fasttext-langdetect 
nltk 
python-docx 
PyPDF2 
olefile 
tqdm 
opencc-python-reimplemented 
python-multipart
openpyxl
anyio
python-dateutil
pymupdf
aiohttp
tiktoken
openai